viper tableview

+++ iOSCourse/Helpers/Const/MeasureConst.swift
@@ -1,20 +1,22 @@
 //
 
-enum ViewIndent: Int {
+import UIKit
+
+enum ViewIndent: CGFloat {
     case tiny = 4
     case small = 8
     case normal = 16
     case medium = 32
 }
 
-enum ViewSize: Int {
+enum ViewSize: CGFloat {
     case tiny = 16
     case small = 32
     case normal = 48
     case medium = 56
 }
 
-enum FontSize: Int {
+enum FontSize: CGFloat {
     case small = 17
     case normal = 24
     case medium = 32
